+++ package.json
@@ -4,10 +4,11 @@
     "description": "",
     "private": true,
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
+        "test": "jest tests/Tasks.test.js --coverage",
         "start": "webpack serve --open",
         "server": "node server.js",
-        "build": "webpack --config prod.config.js"
+        "build": "webpack",
+        "deploy": "gh-pages -d dist"
     },
     "repository": {
         "type": "git",
@@ -22,10 +23,12 @@
     "homepage": "https://github.com/ciraganenicole/To-do-list#readme",
     "dependencies": {
         "express": "^4.17.3",
+        "gh-pages": "^3.2.3",
         "webpack": "^5.70.0",
         "webpack-dev-middleware": "^5.3.1"
     },
     "devDependencies": {
+        "@babel/plugin-transform-modules-commonjs": "^7.17.7",
         "babel-eslint": "^10.1.0",
         "css-loader": "^6.7.1",
         "eslint": "^7.32.0",
@@ -33,6 +36,7 @@
         "eslint-plugin-import": "^2.25.4",
         "hint": "^6.1.10",
         "html-webpack-plugin": "^5.5.0",
+        "jest": "^27.5.1",
         "style-loader": "^3.3.1",
         "stylelint": "^13.13.1",
         "stylelint-config-standard": "^21.0.0",
