+++ webpack.config.js
@@ -2,33 +2,32 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-    mode: 'development',
-    entry: {
-        index: './src/index.js',
-        hamburger: './modules/hamburger.js',
-        Tasks: './modules/Tasks.js',
-        TasksList: './modules/TasksList.js',
-    },
-    devtool: 'inline-source-map',
-    devServer: {
-        static: './dist',
-    },
-    plugins: [
-        new HtmlWebpackPlugin({
-            title: 'Development',
-            template: './src/index.html',
-        }),
-    ],
-    output: {
-        filename: '[name].bundle.js',
-        path: path.resolve(__dirname, 'dist'),
-        clean: true,
-        publicPath: '/',
-    },
-    module: {
-        rules: [{
-            test: /\.css$/i,
-            use: ['style-loader', 'css-loader'],
-        }],
-    },
+  mode: 'development',
+  entry: {
+    index: './src/index.js',
+    Tasks: './modules/Tasks.js',
+    TasksList: './modules/TasksList.js',
+  },
+  devtool: 'inline-source-map',
+  devServer: {
+    static: './dist',
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      title: 'Development',
+      template: './src/index.html',
+    }),
+  ],
+  output: {
+    filename: '[name].bundle.js',
+    path: path.resolve(__dirname, 'dist'),
+    clean: true,
+    publicPath: '/',
+  },
+  module: {
+    rules: [{
+      test: /\.css$/i,
+      use: ['style-loader', 'css-loader'],
+    }],
+  },
 };
